+++ ConsoleApplication1/ConsoleApplication1.cpp
@@ -15,72 +15,88 @@ using namespace std;
 struct znachenia
 {
     vector<double> massiv;
+    double nachalo;
 };
 
 struct pipe
 {
-    double L, dx, dt, V;
+    double L, dx, dt, V, N;
 };
 
-int grid_size = 101;
-int T = 7000; //Заданный период времени 
+int grid_size = 101;//размер сетки
+int T = 3000; //Заданный период времени 
+
+
+
+void rashet(pipe myPipe, double parametr, vector<double>& current_layer, vector<double>& previous_layer) {
+
+    // Суть - смещение предыдущего слоя и запись граничного условия
+    for (size_t i = 1; i < current_layer.size(); i++)
+    {
+        current_layer[i] = previous_layer[i - 1];
+    }
+    current_layer[0] = parametr;
+
+}
+
+
+
+void excel(pipe myPipe, ring_buffer_t<vector<vector<double>>>& buffer, int i) {
+    vector<vector<double>>& current_layer = buffer.current();
+    if (i == 0) {
+        ofstream outFile("plotnostsera.csv");
+        outFile << "время,координиата,плотность, сера" << "\n";
+        // Записать значения текущего слоя в файл
+        for (size_t j = 0; j < current_layer[0].size(); j++) {
+            outFile << i * myPipe.dt << "," << j * myPipe.dx << "," << current_layer[0][j] << "," << current_layer[1][j] << "\n";
+        }
+        outFile.close();
+    }
+    else {
+        ofstream outFile("plotnostsera.csv", ios::app);
+        // Записать значения текущего слоя в файл
+        for (size_t j = 0; j < current_layer[0].size(); j++) {
+            outFile << i * myPipe.dt << "," << j * myPipe.dx << "," << current_layer[0][j] << "," << current_layer[1][j] << "\n";
+        }
+        outFile.close();
+    }
+}
+
+
+
 
 int main()
 {
     pipe myPipe;
     myPipe.L = 100000;
-    myPipe.V = 1.5;
+    myPipe.V = 1;
 
-    double ro_n = 850;
-    double ro_k = 860;
+
     myPipe.dx = myPipe.L / (grid_size - 1);
     myPipe.dt = myPipe.dx / myPipe.V; //условие Куранта
-    int N = (T / myPipe.dt);
+    myPipe.N = (T / myPipe.dt);
 
     znachenia ro;
-    ro.massiv = vector<double>(N);
-    ro.massiv = { 850, 870, 860, 860, 860,860, 860, 860, 860, 860 }; //значения плотности на входе трубы
+    ro.massiv = vector<double>(myPipe.N);
+    ro.massiv = { 850,860,890 }; //значения плотности на входе трубы
+    ro.nachalo = 950;
 
     znachenia sera;
-    sera.massiv = vector<double>(N);
-    sera.massiv = { 0.01, 0.02, 0.03, 0.04, 0.05,0.06, 0.07, 0.08, 0.09, 0.01 }; //значения серы на входе трубы
+    sera.massiv = vector<double>(myPipe.N);
+    sera.massiv = { 0.01, 0.02, 0.03 }; //значения серы на входе трубы
 
-    
-}
 
-void rashet(pipe myPipe, znachenia ro, znachenia sera) {
-    vector<double> initial_layer(grid_size, ro.massiv[0]);
-    // initial_layer - слой, значениями из которого проинициализируются все слои буфера
-    ring_buffer_t<vector<double>> buffer(2, initial_layer);
+    vector<double> odin(grid_size);
+    ring_buffer_t<vector<vector<double>>> buffer(2, { odin, odin });
 
-    for (double time = 0; time < T; time += myPipe.dt)
-    {
-        // Получение ссылок на текущий и предыдущий слои буфера
-        vector<double>& previous_layer = buffer.previous();
-        vector<double>& current_layer = buffer.current();
 
-        // Расчет методом характеристик
-        // Суть - смещение предыдущего слоя и запись граничного условия
-        for (size_t i = 1; i < current_layer.size(); i++)
-        {
-            current_layer[i] = previous_layer[i - 1];
-        }
-        current_layer[0] = ro.massiv[0];
+    for (size_t i = 0; i < myPipe.N; i++) {
+        rashet(myPipe, ro.massiv[i], buffer.current()[0], buffer.previous()[0]);
+        rashet(myPipe, sera.massiv[i], buffer.current()[1], buffer.previous()[1]);
 
-        // Слой current_layer на следующем шаге должен стать предыдущим. 
-        // Для этого сместим индекс текущего слоя в буфере на единицу
-        buffer.advance(+1);
-        
+        excel(myPipe, buffer, i);
+        buffer.advance(1);
     }
-}
 
 
-void excel(pipe myPipe, znachenia ro, znachenia sera) {
-    ofstream outFile("ddddddddddddddddd.csv");
-    outFile << "время,координата,плотность" << "\n";
-    vector<double> znachenia(grid_size);
-    for (int j = 0; j < grid_size; j++) {
-        outFile << j * myPipe.dt << "," << j * myPipe.dx << "," << znachenia[j] << "\n";
-    }
-    outFile.close();
 }
