+++ ConsoleApplication1/plotnostsera.csv
@@ -1,0 +1,304 @@
+время,координиата,плотность, сера
+0,0,850,0.01
+0,1000,0,0
+0,2000,0,0
+0,3000,0,0
+0,4000,0,0
+0,5000,0,0
+0,6000,0,0
+0,7000,0,0
+0,8000,0,0
+0,9000,0,0
+0,10000,0,0
+0,11000,0,0
+0,12000,0,0
+0,13000,0,0
+0,14000,0,0
+0,15000,0,0
+0,16000,0,0
+0,17000,0,0
+0,18000,0,0
+0,19000,0,0
+0,20000,0,0
+0,21000,0,0
+0,22000,0,0
+0,23000,0,0
+0,24000,0,0
+0,25000,0,0
+0,26000,0,0
+0,27000,0,0
+0,28000,0,0
+0,29000,0,0
+0,30000,0,0
+0,31000,0,0
+0,32000,0,0
+0,33000,0,0
+0,34000,0,0
+0,35000,0,0
+0,36000,0,0
+0,37000,0,0
+0,38000,0,0
+0,39000,0,0
+0,40000,0,0
+0,41000,0,0
+0,42000,0,0
+0,43000,0,0
+0,44000,0,0
+0,45000,0,0
+0,46000,0,0
+0,47000,0,0
+0,48000,0,0
+0,49000,0,0
+0,50000,0,0
+0,51000,0,0
+0,52000,0,0
+0,53000,0,0
+0,54000,0,0
+0,55000,0,0
+0,56000,0,0
+0,57000,0,0
+0,58000,0,0
+0,59000,0,0
+0,60000,0,0
+0,61000,0,0
+0,62000,0,0
+0,63000,0,0
+0,64000,0,0
+0,65000,0,0
+0,66000,0,0
+0,67000,0,0
+0,68000,0,0
+0,69000,0,0
+0,70000,0,0
+0,71000,0,0
+0,72000,0,0
+0,73000,0,0
+0,74000,0,0
+0,75000,0,0
+0,76000,0,0
+0,77000,0,0
+0,78000,0,0
+0,79000,0,0
+0,80000,0,0
+0,81000,0,0
+0,82000,0,0
+0,83000,0,0
+0,84000,0,0
+0,85000,0,0
+0,86000,0,0
+0,87000,0,0
+0,88000,0,0
+0,89000,0,0
+0,90000,0,0
+0,91000,0,0
+0,92000,0,0
+0,93000,0,0
+0,94000,0,0
+0,95000,0,0
+0,96000,0,0
+0,97000,0,0
+0,98000,0,0
+0,99000,0,0
+0,100000,0,0
+1000,0,860,0.02
+1000,1000,850,0.01
+1000,2000,0,0
+1000,3000,0,0
+1000,4000,0,0
+1000,5000,0,0
+1000,6000,0,0
+1000,7000,0,0
+1000,8000,0,0
+1000,9000,0,0
+1000,10000,0,0
+1000,11000,0,0
+1000,12000,0,0
+1000,13000,0,0
+1000,14000,0,0
+1000,15000,0,0
+1000,16000,0,0
+1000,17000,0,0
+1000,18000,0,0
+1000,19000,0,0
+1000,20000,0,0
+1000,21000,0,0
+1000,22000,0,0
+1000,23000,0,0
+1000,24000,0,0
+1000,25000,0,0
+1000,26000,0,0
+1000,27000,0,0
+1000,28000,0,0
+1000,29000,0,0
+1000,30000,0,0
+1000,31000,0,0
+1000,32000,0,0
+1000,33000,0,0
+1000,34000,0,0
+1000,35000,0,0
+1000,36000,0,0
+1000,37000,0,0
+1000,38000,0,0
+1000,39000,0,0
+1000,40000,0,0
+1000,41000,0,0
+1000,42000,0,0
+1000,43000,0,0
+1000,44000,0,0
+1000,45000,0,0
+1000,46000,0,0
+1000,47000,0,0
+1000,48000,0,0
+1000,49000,0,0
+1000,50000,0,0
+1000,51000,0,0
+1000,52000,0,0
+1000,53000,0,0
+1000,54000,0,0
+1000,55000,0,0
+1000,56000,0,0
+1000,57000,0,0
+1000,58000,0,0
+1000,59000,0,0
+1000,60000,0,0
+1000,61000,0,0
+1000,62000,0,0
+1000,63000,0,0
+1000,64000,0,0
+1000,65000,0,0
+1000,66000,0,0
+1000,67000,0,0
+1000,68000,0,0
+1000,69000,0,0
+1000,70000,0,0
+1000,71000,0,0
+1000,72000,0,0
+1000,73000,0,0
+1000,74000,0,0
+1000,75000,0,0
+1000,76000,0,0
+1000,77000,0,0
+1000,78000,0,0
+1000,79000,0,0
+1000,80000,0,0
+1000,81000,0,0
+1000,82000,0,0
+1000,83000,0,0
+1000,84000,0,0
+1000,85000,0,0
+1000,86000,0,0
+1000,87000,0,0
+1000,88000,0,0
+1000,89000,0,0
+1000,90000,0,0
+1000,91000,0,0
+1000,92000,0,0
+1000,93000,0,0
+1000,94000,0,0
+1000,95000,0,0
+1000,96000,0,0
+1000,97000,0,0
+1000,98000,0,0
+1000,99000,0,0
+1000,100000,0,0
+2000,0,890,0.03
+2000,1000,860,0.02
+2000,2000,850,0.01
+2000,3000,0,0
+2000,4000,0,0
+2000,5000,0,0
+2000,6000,0,0
+2000,7000,0,0
+2000,8000,0,0
+2000,9000,0,0
+2000,10000,0,0
+2000,11000,0,0
+2000,12000,0,0
+2000,13000,0,0
+2000,14000,0,0
+2000,15000,0,0
+2000,16000,0,0
+2000,17000,0,0
+2000,18000,0,0
+2000,19000,0,0
+2000,20000,0,0
+2000,21000,0,0
+2000,22000,0,0
+2000,23000,0,0
+2000,24000,0,0
+2000,25000,0,0
+2000,26000,0,0
+2000,27000,0,0
+2000,28000,0,0
+2000,29000,0,0
+2000,30000,0,0
+2000,31000,0,0
+2000,32000,0,0
+2000,33000,0,0
+2000,34000,0,0
+2000,35000,0,0
+2000,36000,0,0
+2000,37000,0,0
+2000,38000,0,0
+2000,39000,0,0
+2000,40000,0,0
+2000,41000,0,0
+2000,42000,0,0
+2000,43000,0,0
+2000,44000,0,0
+2000,45000,0,0
+2000,46000,0,0
+2000,47000,0,0
+2000,48000,0,0
+2000,49000,0,0
+2000,50000,0,0
+2000,51000,0,0
+2000,52000,0,0
+2000,53000,0,0
+2000,54000,0,0
+2000,55000,0,0
+2000,56000,0,0
+2000,57000,0,0
+2000,58000,0,0
+2000,59000,0,0
+2000,60000,0,0
+2000,61000,0,0
+2000,62000,0,0
+2000,63000,0,0
+2000,64000,0,0
+2000,65000,0,0
+2000,66000,0,0
+2000,67000,0,0
+2000,68000,0,0
+2000,69000,0,0
+2000,70000,0,0
+2000,71000,0,0
+2000,72000,0,0
+2000,73000,0,0
+2000,74000,0,0
+2000,75000,0,0
+2000,76000,0,0
+2000,77000,0,0
+2000,78000,0,0
+2000,79000,0,0
+2000,80000,0,0
+2000,81000,0,0
+2000,82000,0,0
+2000,83000,0,0
+2000,84000,0,0
+2000,85000,0,0
+2000,86000,0,0
+2000,87000,0,0
+2000,88000,0,0
+2000,89000,0,0
+2000,90000,0,0
+2000,91000,0,0
+2000,92000,0,0
+2000,93000,0,0
+2000,94000,0,0
+2000,95000,0,0
+2000,96000,0,0
+2000,97000,0,0
+2000,98000,0,0
+2000,99000,0,0
+2000,100000,0,0
